Move CheckpointStore trait into its own file
Extracted CheckpointStore, ResultCheckpoint, and ResultSequence from
result_index.rs into checkpoint_store.rs for better separation of concerns.
