+++ core/src/interface/mod.rs
@@ -16,6 +16,7 @@ mod element_index;
 mod future_queue;
 mod index_backend;
 mod query_clock;
+mod checkpoint_store;
 mod result_index;
 mod session_control;
 mod source_middleware;
@@ -37,12 +38,12 @@ pub use index_backend::IndexSet;
 pub use query_clock::QueryClock;
 pub use result_index::AccumulatorIndex;
 pub use result_index::LazySortedSetStore;
-pub use result_index::ResultCheckpoint;
 pub use result_index::ResultIndex;
 pub use result_index::ResultKey;
 pub use result_index::ResultOwner;
-pub use result_index::ResultSequence;
-pub use result_index::CheckpointStore;
+pub use checkpoint_store::CheckpointStore;
+pub use checkpoint_store::ResultCheckpoint;
+pub use checkpoint_store::ResultSequence;
 pub use session_control::NoOpSessionControl;
 pub use session_control::SessionControl;
 pub use session_control::SessionGuard;
